+++ packages/dockest/src/index.ts
@@ -1,7 +1,7 @@
 import { DockestConfig, DockestOpts, DockestService } from './@types'
-import { BaseError, ConfigurationError } from './_errors'
-import { Logger } from './_logger'
 import { MINIMUM_JEST_VERSION } from './constants'
+import { BaseError, ConfigurationError } from './errors'
+import { Logger } from './logger'
 import { bootstrap } from './run/bootstrap'
 import { debugMode } from './run/debugMode'
 import { createLogWriter } from './run/log-writer'
