+++ packages/dockest/src/run/debugMode.ts
@@ -1,5 +1,5 @@
 import { DockestConfig } from '../@types'
-import { Logger } from '../_logger'
+import { Logger } from '../logger'
 import { sleep } from '../utils/sleep'
 
 export const debugMode = async ({
