+++ packages/dockest/src/run/teardown.ts
@@ -1,7 +1,7 @@
 import { LogWriter } from './log-writer'
 import { DockestConfig } from '../@types'
-import { DockestError } from '../_errors'
-import { Logger } from '../_logger'
+import { DockestError } from '../errors'
+import { Logger } from '../logger'
 import { leaveBridgeNetwork } from '../utils/network/leaveBridgeNetwork'
 import { removeBridgeNetwork } from '../utils/network/removeBridgeNetwork'
 import { teardownSingle } from '../utils/teardownSingle'
