+++ packages/dockest/src/run/waitForServices/index.ts
@@ -6,8 +6,8 @@ import { resolveContainerId } from './resolveContainerId'
 import { runReadinessCheck } from './runReadinessCheck'
 import { runRunnerCommands } from './runRunnerCommands'
 import { DockestConfig, Runner } from '../../@types'
-import { DockestError } from '../../_errors'
 import { DOCKEST_HOST_ADDRESS } from '../../constants'
+import { DockestError } from '../../errors'
 import { bridgeNetworkExists } from '../../utils/network/bridgeNetworkExists'
 import { createBridgeNetwork } from '../../utils/network/createBridgeNetwork'
 import { joinBridgeNetwork } from '../../utils/network/joinBridgeNetwork'
