+++ packages/dockest/src/test-helper/index.ts
@@ -1,6 +1,6 @@
 import { DockerComposeFile, TestRunModeType } from '../@types'
-import { DockestError } from '../_errors'
 import { DEFAULT_HOST_NAME, DOCKEST_ATTACH_TO_PROCESS, DOCKEST_HOST_ADDRESS } from '../constants'
+import { DockestError } from '../errors'
 import { getRunMode as _getRunMode } from '../utils/getRunMode'
 import { selectPortMapping } from '../utils/selectPortMapping'
