+++ packages/dockest/src/utils/execaWrapper.ts
@@ -1,7 +1,7 @@
 import execa, { SyncOptions } from 'execa' // eslint-disable-line import/default
 import { trim } from './trim'
 import { Runner } from '../@types'
-import { Logger } from '../_logger'
+import { Logger } from '../logger'
 
 interface Opts {
   runner?: Runner
